pyproject.toml
[tool.poetry]
name = "ai-video-api"
version = "0.1.0"
description = "AI视频生成API封装"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "video_generation"}]
license = "MIT"
repository = "https://github.com/yourusername/ai-video-api"
keywords = ["ai", "video", "generation", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
requests = "^2.31.0"
pydantic = "^2.5.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 